+++ docs/changelog.md
@@ -33,3 +33,4 @@
 - Terminado a aula 01 - Consumindo API - Criando frontend sem backend
 - Terminado a aula 02 - Consumindo API - Configurando o MirageJS
 - Terminado a aula 03 - Consumindo API - Configurando o Axios
+- Terminado a aula 01 - Modal & Forms - Configurando o modal de criação